Refactor query builder concepts out to separate files
