+++ src/realm/CMakeLists.txt
@@ -133,13 +133,23 @@ set(REALM_INSTALL_GENERAL_HEADERS
 ) # REALM_INSTALL_GENERAL_HEADERS
 
 set(REALM_PARSER_SOURCES
+    parser/collection_operator_expression.cpp
+    parser/expression_container.cpp
     parser/parser.cpp
+    parser/parser_utils.cpp
+    parser/property_expression.cpp
     parser/query_builder.cpp
+    parser/value_expression.cpp
 ) # REALM_PARSER_SOURCES
 
 set(REALM_PARSER_HEADERS
+    parser/collection_operator_expression.hpp
+    parser/expression_container.hpp
     parser/parser.hpp
+    parser/parser_utils.hpp
+    parser/property_expression.hpp
     parser/query_builder.hpp
+    parser/value_expression.hpp
 ) # REALM_PARSER_HEADERS
 
 set(REALM_INSTALL_IMPL_HEADERS
