+++ src/realm/parser/query_builder.cpp
@@ -19,15 +19,14 @@
 #include "query_builder.hpp"
 
 #include "parser.hpp"
+#include "parser_utils.hpp"
+#include "expression_container.hpp"
 
 #include <realm.hpp>
 #include <realm/query_expression.hpp>
 #include <realm/column_type.hpp>
 #include <realm/utilities.hpp>
-#include <realm/util/base64.hpp>
 
-#include <functional>
-#include <time.h>
 #include <sstream>
 
 using namespace realm;
@@ -35,565 +34,6 @@ using namespace parser;
 using namespace query_builder;
 
 namespace {
-template<typename T>
-T stot(std::string const& s) {
-    std::istringstream iss(s);
-    T value;
-    iss >> value;
-    if (iss.fail()) {
-        throw std::invalid_argument(util::format("Cannot convert string '%1'", s));
-    }
-    return value;
-}
-
-Timestamp get_timestamp_if_valid(int64_t seconds, int32_t nanoseconds) {
-    const bool both_non_negative = seconds >= 0 && nanoseconds >= 0;
-    const bool both_non_positive = seconds <= 0 && nanoseconds <= 0;
-    if (both_non_negative || both_non_positive) {
-        return Timestamp(seconds, nanoseconds);
-    }
-    throw std::runtime_error("Invalid timestamp format");
-}
-
-Timestamp from_timestamp_values(std::vector<std::string> const& time_inputs) {
-
-    if (time_inputs.size() == 2) {
-        // internal format seconds, nanoseconds
-        int64_t seconds = stot<int64_t>(time_inputs[0]);
-        int32_t nanoseconds = stot<int32_t>(time_inputs[1]);
-        return get_timestamp_if_valid(seconds, nanoseconds);
-    }
-    else if (time_inputs.size() == 6 || time_inputs.size() == 7) {
-        // readable format YYYY-MM-DD-HH:MM:SS:NANOS nanos optional
-        tm created{};
-        created.tm_year = stot<int>(time_inputs[0]) - 1900; // epoch offset (see man mktime)
-        created.tm_mon = stot<int>(time_inputs[1]) - 1; // converts from 1-12 to 0-11
-        created.tm_mday = stot<int>(time_inputs[2]);
-        created.tm_hour = stot<int>(time_inputs[3]);
-        created.tm_min = stot<int>(time_inputs[4]);
-        created.tm_sec = stot<int>(time_inputs[5]);
-
-        if (created.tm_year < 0) {
-            // platform timegm functions do not throw errors, they return -1 which is also a valid time
-            throw std::logic_error("Conversion of dates before 1900 is not supported.");
-        }
-
-        int64_t seconds = platform_timegm(created); // UTC time
-        int32_t nanoseconds = 0;
-        if (time_inputs.size() == 7) {
-            nanoseconds = stot<int32_t>(time_inputs[6]);
-            if (nanoseconds < 0) {
-                throw std::logic_error("The nanoseconds of a Timestamp cannot be negative.");
-            }
-            if (seconds < 0) { // seconds determines the sign of the nanoseconds part
-                nanoseconds *= -1;
-            }
-        }
-        return get_timestamp_if_valid(seconds, nanoseconds);
-    }
-    throw std::runtime_error("Unexpected timestamp format.");
-}
-
-StringData from_base64(const std::string& input, util::StringBuffer& decode_buffer)
-{
-    // expects wrapper tokens B64"..."
-    if (input.size() < 5
-        || !(input[0] == 'B' || input[0] == 'b')
-        || input[1] != '6' || input[2] != '4'
-        || input[3] != '"' || input[input.size() - 1] != '"') {
-        throw std::runtime_error("Unexpected base64 format");
-    }
-    const size_t encoded_size = input.size() - 5;
-    size_t buffer_size = util::base64_decoded_size(encoded_size);
-    decode_buffer.resize(buffer_size);
-    StringData window(input.data() + 4, encoded_size);
-    util::Optional<size_t> decoded_size = util::base64_decode(window, decode_buffer.data(), buffer_size);
-    if (!decoded_size) {
-        throw std::runtime_error("Invalid base64 value");
-    }
-    REALM_ASSERT_DEBUG_EX(*decoded_size <= encoded_size, *decoded_size, encoded_size);
-    decode_buffer.resize(*decoded_size); // truncate
-    StringData output_window(decode_buffer.data(), decode_buffer.size());
-    return output_window;
-}
-
-// check a precondition and throw an exception if it is not met
-// this should be used iff the condition being false indicates a bug in the caller
-// of the function checking its preconditions
-#define precondition(condition, message) if (!REALM_LIKELY(condition)) { throw std::logic_error(message); }
-
-using KeyPath = std::vector<std::string>;
-KeyPath key_path_from_string(const std::string &s) {
-    std::stringstream ss(s);
-    std::string item;
-    KeyPath key_path;
-    while (std::getline(ss, item, '.')) {
-        key_path.push_back(item);
-    }
-    return key_path;
-}
-
-StringData get_printable_table_name(const Table& table)
-{
-    StringData name = table.get_name();
-    // the "class_" prefix is an implementation detail of the object store that shouldn't be exposed to users
-    static const std::string prefix = "class_";
-    if (name.size() > prefix.size() && strncmp(name.data(), prefix.data(), prefix.size()) == 0) {
-        name = StringData(name.data() + prefix.size(), name.size() - prefix.size());
-    }
-    return name;
-}
-
-struct ValueExpression
-{
-    const parser::Expression* value;
-    Arguments* arguments;
-    std::function<Table *()> table_getter;
-
-    ValueExpression(Query& query, Arguments* args, const parser::Expression* v)
-    : value(v)
-    , arguments(args) {
-        table_getter = [&] {
-            auto& tbl = query.get_table();
-            return tbl.get();
-        };
-    }
-    ValueExpression& operator=(const ValueExpression& other)
-    {
-        value = other.value;
-        arguments = other.arguments;
-        table_getter = other.table_getter;
-        return *this;
-    }
-    bool is_null()
-    {
-        if (value->type == parser::Expression::Type::Null) {
-            return true;
-        }
-        else if (value->type == parser::Expression::Type::Argument) {
-            return arguments->is_argument_null(stot<int>(value->s));
-        }
-        return false;
-    }
-
-    template <typename RetType>
-    RetType value_of_type_for_query();
-};
-
-
-template <>
-Timestamp ValueExpression::value_of_type_for_query<Timestamp>()
-{
-    if (value->type == parser::Expression::Type::Argument) {
-        return arguments->timestamp_for_argument(stot<int>(value->s));
-    } else if (value->type == parser::Expression::Type::Timestamp) {
-        return from_timestamp_values(value->time_inputs);
-    } else if (value->type == parser::Expression::Type::Null) {
-        return Timestamp(realm::null());
-    }
-    throw std::logic_error("Attempting to compare Timestamp property to a non-Timestamp value");
-}
-
-template <>
-bool ValueExpression::value_of_type_for_query<bool>()
-{
-    if (value->type == parser::Expression::Type::Argument) {
-        return arguments->bool_for_argument(stot<int>(value->s));
-    }
-    if (value->type != parser::Expression::Type::True && value->type != parser::Expression::Type::False) {
-        if (value->type == parser::Expression::Type::Number) {
-            // As a special exception we can handle 0 and 1.
-            // Our bool values are actually stored as integers {0, 1}
-            int64_t number_value = stot<int64_t>(value->s);
-            if (number_value == 0) {
-                return false;
-            }
-            else if (number_value == 1) {
-                return true;
-            }
-        }
-        throw std::logic_error("Attempting to compare bool property to a non-bool value");
-    }
-    return value->type == parser::Expression::Type::True;
-}
-
-template <>
-Double ValueExpression::value_of_type_for_query<Double>()
-{
-    if (value->type == parser::Expression::Type::Argument) {
-        return arguments->double_for_argument(stot<int>(value->s));
-    }
-    return stot<double>(value->s);
-}
-
-template <>
-Float ValueExpression::value_of_type_for_query<Float>()
-{
-    if (value->type == parser::Expression::Type::Argument) {
-        return arguments->float_for_argument(stot<int>(value->s));
-    }
-    return stot<float>(value->s);
-}
-
-template <>
-Int ValueExpression::value_of_type_for_query<Int>()
-{
-    if (value->type == parser::Expression::Type::Argument) {
-        return arguments->long_for_argument(stot<int>(value->s));
-    }
-    return stot<long long>(value->s);
-}
-
-template <>
-StringData ValueExpression::value_of_type_for_query<StringData>()
-{
-    if (value->type == parser::Expression::Type::Argument) {
-        return arguments->string_for_argument(stot<int>(value->s));
-    }
-    else if (value->type == parser::Expression::Type::String) {
-        return value->s;
-    }
-    else if (value->type == parser::Expression::Type::Base64) {
-        // the return value points to data in the lifetime of args
-        return from_base64(value->s, arguments->buffer_space);
-    }
-    throw std::logic_error("Attempting to compare String property to a non-String value");
-}
-
-template <>
-BinaryData ValueExpression::value_of_type_for_query<BinaryData>()
-{
-    if (value->type == parser::Expression::Type::Argument) {
-        return arguments->binary_for_argument(stot<int>(value->s));
-    }
-    else if (value->type == parser::Expression::Type::String) {
-        return BinaryData(value->s);
-    }
-    else if (value->type == parser::Expression::Type::Base64) {
-        StringData converted = from_base64(value->s, arguments->buffer_space);
-        // returning a pointer to data in the lifetime of args
-        return BinaryData(converted.data(), converted.size());
-    }
-    throw std::logic_error("Binary properties must be compared against a binary argument.");
-}
-
-struct PropertyExpression
-{
-    std::vector<size_t> indexes;
-    size_t col_ndx;
-    DataType col_type;
-    Query &query;
-
-    PropertyExpression(Query &q, const std::string &key_path_string)
-    : query(q)
-    {
-        KeyPath key_path = key_path_from_string(key_path_string);
-        TableRef cur_table = query.get_table();
-        for (size_t index = 0; index < key_path.size(); index++) {
-            size_t cur_col_ndx = cur_table->get_column_index(key_path[index]);
-
-            StringData object_name = get_printable_table_name(*cur_table);
-
-            precondition(cur_col_ndx != realm::not_found,
-                         util::format("No property '%1' on object of type '%2'", key_path[index], object_name));
-
-            DataType cur_col_type = cur_table->get_column_type(cur_col_ndx);
-            if (index != key_path.size() - 1) {
-                precondition(cur_col_type == type_Link || cur_col_type == type_LinkList,
-                             util::format("Property '%1' is not a link in object of type '%2'", key_path[index], object_name));
-                indexes.push_back(cur_col_ndx);
-                cur_table = cur_table->get_link_target(cur_col_ndx);
-            }
-            else {
-                col_ndx = cur_col_ndx;
-                col_type = cur_col_type;
-            }
-        }
-    }
-
-    Table* table_getter() const
-    {
-        auto& tbl = query.get_table();
-        for (size_t col : indexes) {
-            tbl->link(col); // mutates m_link_chain on table
-        }
-        return tbl.get();
-    }
-
-    template <typename RetType>
-    auto value_of_type_for_query() const
-    {
-        return this->table_getter()->template column<RetType>(this->col_ndx);
-    }
-};
-template <typename T>
-const char* type_to_str()
-{
-    return typeid(T).name();
-}
-template <>
-const char* type_to_str<bool>()
-{
-    return "Bool";
-}
-template <>
-const char* type_to_str<Int>()
-{
-    return "Int";
-}
-template <>
-const char* type_to_str<Float>()
-{
-    return "Float";
-}
-template <>
-const char* type_to_str<Double>()
-{
-    return "Double";
-}
-template <>
-const char* type_to_str<String>()
-{
-    return "String";
-}
-template <>
-const char* type_to_str<Binary>()
-{
-    return "Binary";
-}
-template <>
-const char* type_to_str<OldDateTime>()
-{
-    return "DateTime";
-}
-template <>
-const char* type_to_str<Timestamp>()
-{
-    return "Timestamp";
-}
-template <>
-const char* type_to_str<Table>()
-{
-    return "Table";
-}
-template <>
-const char* type_to_str<Mixed>()
-{
-    return "Mixed";
-}
-template <>
-const char* type_to_str<Link>()
-{
-    return "Link";
-}
-
-const char* data_type_to_str(DataType type)
-{
-    switch (type) {
-        case type_Int:
-            return "Int";
-        case type_Bool:
-            return "Bool";
-        case type_Float:
-            return "Float";
-        case type_Double:
-            return "Double";
-        case type_String:
-            return "String";
-        case type_Binary:
-            return "Binary";
-        case type_OldDateTime:
-            return "DateTime";
-        case type_Timestamp:
-            return "Timestamp";
-        case type_Table:
-            return "Table";
-        case type_Mixed:
-            return "Mixed";
-        case type_Link:
-            return "Link";
-        case type_LinkList:
-            return "LinkList";
-    }
-    return "type_Unknown"; // LCOV_EXCL_LINE
-}
-
-const char* collection_operator_to_str(parser::Expression::KeyPathOp op)
-{
-    switch (op) {
-        case parser::Expression::KeyPathOp::None:
-            return "NONE";
-        case parser::Expression::KeyPathOp::Min:
-            return "@min";
-        case parser::Expression::KeyPathOp::Max:
-            return "@max";
-        case parser::Expression::KeyPathOp::Sum:
-            return "@sum";
-        case parser::Expression::KeyPathOp::Avg:
-            return "@avg";
-        case parser::Expression::KeyPathOp::SizeString:
-            return "@size";
-        case parser::Expression::KeyPathOp::SizeBinary:
-            return "@size";
-        case parser::Expression::KeyPathOp::Count:
-            return "@count";
-    }
-    return "";
-}
-
-template <typename RetType, parser::Expression::KeyPathOp AggOpType, class Enable = void>
-struct CollectionOperatorGetter;
-
-template <parser::Expression::KeyPathOp OpType>
-struct CollectionOperatorExpression
-{
-    static constexpr parser::Expression::KeyPathOp operation_type = OpType;
-    std::function<Table *()> table_getter;
-    PropertyExpression pe;
-    size_t post_link_col_ndx;
-    DataType post_link_col_type;
-    CollectionOperatorExpression(PropertyExpression&& exp, std::string suffix_path)
-    : pe(std::move(exp))
-    , post_link_col_ndx(realm::not_found)
-    {
-        table_getter = std::bind(&PropertyExpression::table_getter, pe);
-
-        const bool requires_suffix_path = !(OpType == parser::Expression::KeyPathOp::SizeString
-                                            || OpType == parser::Expression::KeyPathOp::SizeBinary
-                                            || OpType == parser::Expression::KeyPathOp::Count);
-
-        if (requires_suffix_path) {
-            Table* pre_link_table = pe.table_getter();
-            REALM_ASSERT(pre_link_table);
-            StringData list_property_name = pre_link_table->get_column_name(pe.col_ndx);
-            precondition(pe.col_type == type_LinkList,
-                         util::format("The '%1' operation must be used on a list property, but '%2' is not a list",
-                                      collection_operator_to_str(OpType), list_property_name));
-
-            TableRef post_link_table = pe.table_getter()->get_link_target(pe.col_ndx);
-            REALM_ASSERT(post_link_table);
-            StringData printable_post_link_table_name = get_printable_table_name(*post_link_table);
-
-            KeyPath suffix_key_path = key_path_from_string(suffix_path);
-            precondition(suffix_path.size() > 0 && suffix_key_path.size() > 0,
-                         util::format("A property from object '%1' must be provided to perform operation '%2'",
-                                      printable_post_link_table_name, collection_operator_to_str(OpType)));
-
-            precondition(suffix_key_path.size() == 1,
-                         util::format("Unable to use '%1' because collection aggreate operations are only supported "
-                                      "for direct properties at this time", suffix_path));
-
-            post_link_col_ndx = pe.table_getter()->get_link_target(pe.col_ndx)->get_column_index(suffix_key_path[0]);
-
-            precondition(post_link_col_ndx != realm::not_found,
-                         util::format("No property '%1' on object of type '%2'",
-                                      suffix_path, printable_post_link_table_name));
-            post_link_col_type = pe.table_getter()->get_link_target(pe.col_ndx)->get_column_type(post_link_col_ndx);
-        }
-        else {  // !requires_suffix_path
-            post_link_col_type = pe.col_type;
-
-            precondition(suffix_path.empty(),
-                         util::format("An extraneous property '%1' was found for operation '%2'",
-                                      suffix_path, collection_operator_to_str(OpType)));
-        }
-    }
-
-    template <typename T>
-    auto value_of_type_for_query() const
-    {
-        return CollectionOperatorGetter<T, OpType>::convert(*this);
-    }
-};
-
-// Certain operations are disabled for some types (eg. a sum of timestamps is invalid).
-// The operations that are supported have a specialisation with std::enable_if for that type below
-// any type/operation combination that is not specialised will get the runtime error from the following
-// default implementation. The return type is just a dummy to make things compile.
-template <typename RetType, parser::Expression::KeyPathOp AggOpType, class Enable>
-struct CollectionOperatorGetter {
-    static Columns<RetType> convert(const CollectionOperatorExpression<AggOpType>& op) {
-        throw std::runtime_error(util::format("Predicate error: comparison of type '%1' with result of '%2' is not supported.",
-                                              type_to_str<RetType>(),
-                                              collection_operator_to_str(op.operation_type)));
-    }
-};
-
-template <typename RetType>
-struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Min,
-    typename std::enable_if_t<
-        std::is_same<RetType, Int>::value ||
-        std::is_same<RetType, Float>::value ||
-        std::is_same<RetType, Double>::value
-    > >{
-    static SubColumnAggregate<RetType, aggregate_operations::Minimum<RetType> > convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::Min>& expr)
-    {
-        return expr.table_getter()->template column<Link>(expr.pe.col_ndx).template column<RetType>(expr.post_link_col_ndx).min();
-    }
-};
-
-template <typename RetType>
-struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Max,
-    typename std::enable_if_t<
-        std::is_same<RetType, Int>::value ||
-        std::is_same<RetType, Float>::value ||
-        std::is_same<RetType, Double>::value
-    > >{
-    static SubColumnAggregate<RetType, aggregate_operations::Maximum<RetType> > convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::Max>& expr)
-    {
-        return expr.table_getter()->template column<Link>(expr.pe.col_ndx).template column<RetType>(expr.post_link_col_ndx).max();
-    }
-};
-
-template <typename RetType>
-struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Sum,
-    typename std::enable_if_t<
-        std::is_same<RetType, Int>::value ||
-        std::is_same<RetType, Float>::value ||
-        std::is_same<RetType, Double>::value
-    > >{
-    static SubColumnAggregate<RetType, aggregate_operations::Sum<RetType> > convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::Sum>& expr)
-    {
-        return expr.table_getter()->template column<Link>(expr.pe.col_ndx).template column<RetType>(expr.post_link_col_ndx).sum();
-    }
-};
-
-template <typename RetType>
-struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Avg,
-    typename std::enable_if_t<
-        std::is_same<RetType, Int>::value ||
-        std::is_same<RetType, Float>::value ||
-        std::is_same<RetType, Double>::value
-    > >{
-    static SubColumnAggregate<RetType, aggregate_operations::Average<RetType> > convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::Avg>& expr)
-    {
-        return expr.table_getter()->template column<Link>(expr.pe.col_ndx).template column<RetType>(expr.post_link_col_ndx).average();
-    }
-};
-
-template <>
-struct CollectionOperatorGetter<Int, parser::Expression::KeyPathOp::Count>{
-    static LinkCount convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::Count>& expr)
-    {
-        return expr.table_getter()->template column<Link>(expr.pe.col_ndx).count();
-    }
-};
-
-template <>
-struct CollectionOperatorGetter<Int, parser::Expression::KeyPathOp::SizeString>{
-    static SizeOperator<Size<String> > convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::SizeString>& expr)
-    {
-        return expr.table_getter()->template column<String>(expr.pe.col_ndx).size();
-    }
-};
-
-template <>
-struct CollectionOperatorGetter<Int, parser::Expression::KeyPathOp::SizeBinary>{
-    static SizeOperator<Size<Binary> > convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::SizeBinary>& expr)
-    {
-        return expr.table_getter()->template column<Binary>(expr.pe.col_ndx).size();
-    }
-};
 
 
 template<typename T, parser::Expression::KeyPathOp OpType>
@@ -689,7 +129,7 @@ void do_add_null_comparison_to_query(Query &query, Predicate::Comparison cmp, co
             do_add_null_comparison_to_query<Link>(query, cmp.op, expr);
             break;
         default:
-            throw std::logic_error(util::format("Object type '%1' not supported", data_type_to_str(type)));
+            throw std::logic_error(util::format("Object type '%1' not supported", util::data_type_to_str(type)));
     }
 }
 
@@ -737,39 +177,6 @@ void add_bool_constraint_to_query(Query &query, Predicate::Operator operatorType
             throw std::logic_error("Unsupported operator for numeric queries.");
     }
 }
-
-void add_string_constraint_to_query(Query &,
-                                    Predicate::Comparison,
-                                    SizeOperator<Size<StringData>> &&,
-                                    StringData &&) {
-    throw std::logic_error("Comparing the size of a string with a string literal is an unsupported operation.");
-}
-void add_string_constraint_to_query(Query &,
-                                    Predicate::Comparison,
-                                    SizeOperator<Size<StringData>> &&,
-                                    Columns<String> &&) {
-    throw std::logic_error("Comparing the size of a string with a string property is an unsupported operation.");
-}
-void add_string_constraint_to_query(Query &,
-                                    Predicate::Comparison,
-                                    StringData &&,
-                                    SizeOperator<Size<StringData>> &&) {
-    throw std::logic_error("Comparing the size of a string with a string literal is an unsupported operation.");
-}
-void add_string_constraint_to_query(Query &,
-                                    Predicate::Comparison,
-                                    Columns<String> &&,
-                                    SizeOperator<Size<StringData>> &&) {
-    throw std::logic_error("Comparing the size of a string with a string property is an unsupported operation.");
-}
-void add_string_constraint_to_query(Query &query,
-                                    Predicate::Comparison cmp,
-                                    SizeOperator<Size<StringData>> &&lhs,
-                                    SizeOperator<Size<StringData>> &&rhs) {
-    // comparing the result of two size operations is a numeric constraint
-    add_numeric_constraint_to_query(query, cmp.op, std::move(lhs), std::move(rhs));
-}
-
 
 void add_string_constraint_to_query(Query &query,
                                     Predicate::Comparison cmp,
@@ -848,38 +255,6 @@ void add_string_constraint_to_query(realm::Query &query,
     }
 }
 
-void add_binary_constraint_to_query(Query &,
-                                    Predicate::Operator,
-                                    SizeOperator<Size<BinaryData>> &&,
-                                    BinaryData &&) {
-    throw std::logic_error("Comparing the size of a binary with a binary literal is an unsupported operation.");
-}
-void add_binary_constraint_to_query(Query &,
-                                    Predicate::Operator,
-                                    SizeOperator<Size<BinaryData>> &&,
-                                    Columns<BinaryData> &&) {
-    throw std::logic_error("Comparing the size of a binary with a binary property is an unsupported operation.");
-}
-void add_binary_constraint_to_query(Query &,
-                                    Predicate::Operator,
-                                    BinaryData &&,
-                                    SizeOperator<Size<BinaryData>> &&) {
-    throw std::logic_error("Comparing the size of a binary with a binary literal is an unsupported operation.");
-}
-void add_binary_constraint_to_query(Query &,
-                                    Predicate::Operator,
-                                    Columns<BinaryData> &&,
-                                    SizeOperator<Size<BinaryData>> &&) {
-    throw std::logic_error("Comparing the size of a binary with a binary property is an unsupported operation.");
-}
-void add_binary_constraint_to_query(Query &query,
-                                    Predicate::Operator op,
-                                    SizeOperator<Size<BinaryData>> &&lhs,
-                                    SizeOperator<Size<BinaryData>> &&rhs) {
-    // comparing the result of two size operations is a numeric constraint
-    add_numeric_constraint_to_query(query, op, std::move(lhs), std::move(rhs));
-}
-
 void add_binary_constraint_to_query(Query &query,
                                     Predicate::Operator op,
                                     Columns<Binary> &&column,
@@ -928,10 +303,21 @@ void add_binary_constraint_to_query(Query &,
     throw std::logic_error("Comparing two binary columns is currently unsupported.");
 }
 
+
+template <typename LHS, typename RHS>
+void add_link_constraint_to_query(realm::Query &,
+                                  Predicate::Operator,
+                                  const LHS &,
+                                  const RHS &) {
+    throw std::runtime_error("Object comparisons are currently only supported between a property and an argument.");
+}
+
+template <>
 void add_link_constraint_to_query(realm::Query &query,
                                   Predicate::Operator op,
                                   const PropertyExpression &prop_expr,
-                                  size_t row_index) {
+                                  const ValueExpression &value_expr) {
+    size_t row_index = value_expr.arguments->object_index_for_argument(stot<int>(value_expr.value->s));
     precondition(prop_expr.indexes.empty(), "KeyPath queries not supported for object comparisons.");
     switch (op) {
         case Predicate::Operator::NotEqual:
@@ -947,43 +333,15 @@ void add_link_constraint_to_query(realm::Query &query,
     }
 }
 
-size_t link_argument(const PropertyExpression&, const PropertyExpression&, Arguments&)
-{
-    throw std::logic_error("Comparing two link properties is not supported at this time.");
+template <>
+void add_link_constraint_to_query(realm::Query &query,
+                                  Predicate::Operator op,
+                                  const ValueExpression &value_expr,
+                                  const PropertyExpression &prop_expr) {
+    // since equality is commutative we can just call the above function to the same effect
+    add_link_constraint_to_query(query, op, prop_expr, value_expr);
 }
 
-size_t link_argument(const PropertyExpression&, const parser::Expression &argExpr, Arguments &args)
-{
-    return args.object_index_for_argument(stot<int>(argExpr.s));
-}
-
-size_t link_argument(const parser::Expression &argExpr, const PropertyExpression&, Arguments &args)
-{
-    return args.object_index_for_argument(stot<int>(argExpr.s));
-}
-
-//size_t link_argument(const CollectionOperatorExpression&, const parser::Expression &, Arguments &)
-//{
-//    throw std::logic_error("Unsupported link comparison");
-//}
-//
-//size_t link_argument(const parser::Expression &, const CollectionOperatorExpression&, Arguments &)
-//{
-//    throw std::logic_error("Unsupported link comparison");
-//}
-//size_t link_argument(const CollectionOperatorExpression&, const PropertyExpression &, Arguments &)
-//{
-//    throw std::logic_error("Unsupported link comparison");
-//}
-//
-//size_t link_argument(const PropertyExpression &, const CollectionOperatorExpression&, Arguments &)
-//{
-//    throw std::logic_error("Unsupported link comparison");
-//}
-//size_t link_argument(const CollectionOperatorExpression &, const CollectionOperatorExpression&, Arguments &)
-//{
-//    throw std::logic_error("Unsupported link comparison");
-//}
 
 template <typename A, typename B>
 void do_add_comparison_to_query(Query &query, Predicate::Comparison cmp, A &lhs, B &rhs, DataType type)
@@ -1026,7 +384,7 @@ void do_add_comparison_to_query(Query &query, Predicate::Comparison cmp, A &lhs,
                                            rhs. template value_of_type_for_query<Binary>());
             break;
         case type_Link:
-            //add_link_constraint_to_query(query, cmp.op, expr, link_argument(lhs, rhs, args));
+            add_link_constraint_to_query(query, cmp.op, lhs, rhs);
             break;
         default:
             throw std::logic_error(util::format("Object type '%1' not supported", data_type_to_str(type)));
@@ -1039,272 +397,105 @@ void do_add_comparison_to_query(Query&, Predicate::Comparison, ValueExpression&,
     throw std::runtime_error("Invalid predicate: comparison between two literals is not supported.");
 }
 
-struct ExpressionContainer
+void add_null_comparison_to_query(Query &query, Predicate::Comparison cmp, ExpressionContainer& exp)
 {
-    enum class ExpressionInternal
-    {
-        exp_Value,
-        exp_Property,
-        exp_OpMin,
-        exp_OpMax,
-        exp_OpSum,
-        exp_OpAvg,
-        exp_OpCount,
-        exp_OpSizeString,
-        exp_OpSizeBinary
-    };
+    switch (exp.type) {
+        case ExpressionContainer::ExpressionInternal::exp_Value:
+            throw std::runtime_error("Unsupported query. A comparison must include at least one keypath.");
+        case ExpressionContainer::ExpressionInternal::exp_Property:
+            do_add_null_comparison_to_query(query, cmp, exp.get_property(), exp.get_property().col_type);
+            break;
+        case ExpressionContainer::ExpressionInternal::exp_OpMin:
+            do_add_null_comparison_to_query(query, cmp, exp.get_min(), exp.get_min().post_link_col_type);
+            break;
+        case ExpressionContainer::ExpressionInternal::exp_OpMax:
+            do_add_null_comparison_to_query(query, cmp, exp.get_max(), exp.get_max().post_link_col_type);
+            break;
+        case ExpressionContainer::ExpressionInternal::exp_OpSum:
+            do_add_null_comparison_to_query(query, cmp, exp.get_sum(), exp.get_sum().post_link_col_type);
+            break;
+        case ExpressionContainer::ExpressionInternal::exp_OpAvg:
+            do_add_null_comparison_to_query(query, cmp, exp.get_avg(), exp.get_avg().post_link_col_type);
+            break;
+        case ExpressionContainer::ExpressionInternal::exp_OpCount:
+            do_add_null_comparison_to_query(query, cmp, exp.get_count(), exp.get_count().post_link_col_type);
+            break;
+        case ExpressionContainer::ExpressionInternal::exp_OpSizeString:
+            do_add_null_comparison_to_query(query, cmp, exp.get_size_string(), exp.get_size_string().post_link_col_type);
+            break;
+        case ExpressionContainer::ExpressionInternal::exp_OpSizeBinary:
+            do_add_null_comparison_to_query(query, cmp, exp.get_size_binary(), exp.get_size_binary().post_link_col_type);
+            break;
+    }
+}
 
-    ExpressionInternal type;
-    util::Any storage;
+template <typename LHS_T>
+void internal_add_comparison_to_query(Query& query, LHS_T& lhs, Predicate::Comparison cmp, ExpressionContainer& rhs, DataType comparison_type)
+{
+    switch (rhs.type) {
+        case ExpressionContainer::ExpressionInternal::exp_Value:
+            do_add_comparison_to_query(query, cmp, lhs, rhs.get_value(), comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_Property:
+            do_add_comparison_to_query(query, cmp, lhs, rhs.get_property(), comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpMin:
+            do_add_comparison_to_query(query, cmp, lhs, rhs.get_min(), comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpMax:
+            do_add_comparison_to_query(query, cmp, lhs, rhs.get_max(), comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpSum:
+            do_add_comparison_to_query(query, cmp, lhs, rhs.get_sum(), comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpAvg:
+            do_add_comparison_to_query(query, cmp, lhs, rhs.get_avg(), comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpCount:
+            do_add_comparison_to_query(query, cmp, lhs, rhs.get_count(), comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpSizeString:
+            do_add_comparison_to_query(query, cmp, lhs, rhs.get_size_string(), comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpSizeBinary:
+            do_add_comparison_to_query(query, cmp, lhs, rhs.get_size_binary(), comparison_type);
+            return;
+    }
+}
 
-    ExpressionContainer(Query& query, const parser::Expression& e, Arguments& args)
-    {
-        if (e.type == parser::Expression::Type::KeyPath) {
-            PropertyExpression pe(query, e.s);
-            switch (e.collection_op) {
-                case parser::Expression::KeyPathOp::Min:
-                    type = ExpressionInternal::exp_OpMin;
-                    storage = CollectionOperatorExpression<parser::Expression::KeyPathOp::Min>(std::move(pe), e.op_suffix);
-                    break;
-                case parser::Expression::KeyPathOp::Max:
-                    type = ExpressionInternal::exp_OpMax;
-                    storage = CollectionOperatorExpression<parser::Expression::KeyPathOp::Max>(std::move(pe), e.op_suffix);
-                    break;
-                case parser::Expression::KeyPathOp::Sum:
-                    type = ExpressionInternal::exp_OpSum;
-                    storage = CollectionOperatorExpression<parser::Expression::KeyPathOp::Sum>(std::move(pe), e.op_suffix);
-                    break;
-                case parser::Expression::KeyPathOp::Avg:
-                    type = ExpressionInternal::exp_OpAvg;
-                    storage = CollectionOperatorExpression<parser::Expression::KeyPathOp::Avg>(std::move(pe), e.op_suffix);
-                    break;
-                case parser::Expression::KeyPathOp::Count:
-                    REALM_FALLTHROUGH;
-                case parser::Expression::KeyPathOp::SizeString:
-                    REALM_FALLTHROUGH;
-                case parser::Expression::KeyPathOp::SizeBinary:
-                    if (pe.col_type == type_LinkList) {
-                        type = ExpressionInternal::exp_OpCount;
-                        storage = CollectionOperatorExpression<parser::Expression::KeyPathOp::Count>(std::move(pe), e.op_suffix);
-                    }
-                    else if (pe.col_type == type_String) {
-                        type = ExpressionInternal::exp_OpSizeString;
-                        storage = CollectionOperatorExpression<parser::Expression::KeyPathOp::SizeString>(std::move(pe), e.op_suffix);
-                    }
-                    else if (pe.col_type == type_Binary) {
-                        type = ExpressionInternal::exp_OpSizeBinary;
-                        storage = CollectionOperatorExpression<parser::Expression::KeyPathOp::SizeBinary>(std::move(pe), e.op_suffix);
-                    }
-                    else {
-                        throw std::runtime_error("Invalid query: @size and @count can only operate on types list, binary, or string");
-                    }
-                    break;
-                case parser::Expression::KeyPathOp::None:
-                    type = ExpressionInternal::exp_Property;
-                    storage = std::move(pe);
-                    break;
-            }
-        }
-        else {
-            type = ExpressionInternal::exp_Value;
-            storage = ValueExpression(query, &args, &e);
-        }
+void add_comparison_to_query(Query &query, ExpressionContainer& lhs, Predicate::Comparison cmp, ExpressionContainer& rhs)
+{
+    DataType comparison_type = lhs.get_comparison_type(rhs);
+    switch (lhs.type) {
+        case ExpressionContainer::ExpressionInternal::exp_Value:
+            internal_add_comparison_to_query(query, lhs.get_value(), cmp, rhs, comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_Property:
+            internal_add_comparison_to_query(query, lhs.get_property(), cmp, rhs, comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpMin:
+            internal_add_comparison_to_query(query, lhs.get_min(), cmp, rhs, comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpMax:
+            internal_add_comparison_to_query(query, lhs.get_max(), cmp, rhs, comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpSum:
+            internal_add_comparison_to_query(query, lhs.get_sum(), cmp, rhs, comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpAvg:
+            internal_add_comparison_to_query(query, lhs.get_avg(), cmp, rhs, comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpCount:
+            internal_add_comparison_to_query(query, lhs.get_count(), cmp, rhs, comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpSizeString:
+            internal_add_comparison_to_query(query, lhs.get_size_string(), cmp, rhs, comparison_type);
+            return;
+        case ExpressionContainer::ExpressionInternal::exp_OpSizeBinary:
+            internal_add_comparison_to_query(query, lhs.get_size_binary(), cmp, rhs, comparison_type);
+            return;
     }
-
-    void add_null_comparison_to_query(Query &query, Predicate::Comparison cmp)
-    {
-
-        switch (type) {
-            case ExpressionInternal::exp_Value:
-                throw std::runtime_error("Unsupported query. A comparison must include at least one keypath.");
-            case ExpressionInternal::exp_Property:
-                do_add_null_comparison_to_query(query, cmp, get_property(), get_property().col_type);
-                break;
-            case ExpressionInternal::exp_OpMin:
-                do_add_null_comparison_to_query(query, cmp, get_min(), get_min().post_link_col_type);
-                break;
-            case ExpressionInternal::exp_OpMax:
-                do_add_null_comparison_to_query(query, cmp, get_max(), get_max().post_link_col_type);
-                break;
-            case ExpressionInternal::exp_OpSum:
-                do_add_null_comparison_to_query(query, cmp, get_sum(), get_sum().post_link_col_type);
-                break;
-            case ExpressionInternal::exp_OpAvg:
-                do_add_null_comparison_to_query(query, cmp, get_avg(), get_avg().post_link_col_type);
-                break;
-            case ExpressionInternal::exp_OpCount:
-                do_add_null_comparison_to_query(query, cmp, get_count(), get_count().post_link_col_type);
-                break;
-            case ExpressionInternal::exp_OpSizeString:
-                do_add_null_comparison_to_query(query, cmp, get_size_string(), get_size_string().post_link_col_type);
-                break;
-            case ExpressionInternal::exp_OpSizeBinary:
-                do_add_null_comparison_to_query(query, cmp, get_size_binary(), get_size_binary().post_link_col_type);
-                break;
-        }
-    }
-
-    PropertyExpression& get_property()
-    {
-        REALM_ASSERT_DEBUG(type == ExpressionInternal::exp_Property);
-        return util::any_cast<PropertyExpression&>(storage);
-    }
-    ValueExpression& get_value()
-    {
-        REALM_ASSERT_DEBUG(type == ExpressionInternal::exp_Value);
-        return util::any_cast<ValueExpression&>(storage);
-    }
-    CollectionOperatorExpression<parser::Expression::KeyPathOp::Min>& get_min()
-    {
-        REALM_ASSERT_DEBUG(type == ExpressionInternal::exp_OpMin);
-        return util::any_cast<CollectionOperatorExpression<parser::Expression::KeyPathOp::Min>&>(storage);
-    }
-    CollectionOperatorExpression<parser::Expression::KeyPathOp::Max>& get_max()
-    {
-        REALM_ASSERT_DEBUG(type == ExpressionInternal::exp_OpMax);
-        return util::any_cast<CollectionOperatorExpression<parser::Expression::KeyPathOp::Max>&>(storage);
-    }
-    CollectionOperatorExpression<parser::Expression::KeyPathOp::Sum>& get_sum()
-    {
-        REALM_ASSERT_DEBUG(type == ExpressionInternal::exp_OpSum);
-        return util::any_cast<CollectionOperatorExpression<parser::Expression::KeyPathOp::Sum>&>(storage);
-    }
-    CollectionOperatorExpression<parser::Expression::KeyPathOp::Avg>& get_avg()
-    {
-        REALM_ASSERT_DEBUG(type == ExpressionInternal::exp_OpAvg);
-        return util::any_cast<CollectionOperatorExpression<parser::Expression::KeyPathOp::Avg>&>(storage);
-    }
-    CollectionOperatorExpression<parser::Expression::KeyPathOp::Count>& get_count()
-    {
-        REALM_ASSERT_DEBUG(type == ExpressionInternal::exp_OpCount);
-        return util::any_cast<CollectionOperatorExpression<parser::Expression::KeyPathOp::Count>&>(storage);
-    }
-    CollectionOperatorExpression<parser::Expression::KeyPathOp::SizeString>& get_size_string()
-    {
-        REALM_ASSERT_DEBUG(type == ExpressionInternal::exp_OpSizeString);
-        return util::any_cast<CollectionOperatorExpression<parser::Expression::KeyPathOp::SizeString>&>(storage);
-    }
-    CollectionOperatorExpression<parser::Expression::KeyPathOp::SizeBinary>& get_size_binary()
-    {
-        REALM_ASSERT_DEBUG(type == ExpressionInternal::exp_OpSizeBinary);
-        return util::any_cast<CollectionOperatorExpression<parser::Expression::KeyPathOp::SizeBinary>&>(storage);
-    }
-
-    DataType get_comparison_type(ExpressionContainer& rhs) {
-        if (type == ExpressionInternal::exp_Property) {
-            return get_property().col_type;
-        } else if (rhs.type == ExpressionInternal::exp_Property) {
-            return rhs.get_property().col_type;
-        } else if (type == ExpressionInternal::exp_OpMin) {
-            return get_min().post_link_col_type;
-        } else if (type == ExpressionInternal::exp_OpMax) {
-            return get_max().post_link_col_type;
-        } else if (type == ExpressionInternal::exp_OpSum) {
-            return get_sum().post_link_col_type;
-        } else if (type == ExpressionInternal::exp_OpAvg) {
-            return get_avg().post_link_col_type;
-        } else if (type == ExpressionInternal::exp_OpCount) {
-            return type_Int;
-        } else if (type == ExpressionInternal::exp_OpSizeString) {
-            return type_Int;
-        } else if (type == ExpressionInternal::exp_OpSizeBinary) {
-            return type_Int;
-        }
-        // rhs checks
-        else if (rhs.type == ExpressionInternal::exp_OpMin) {
-            return rhs.get_min().post_link_col_type;
-        } else if (rhs.type == ExpressionInternal::exp_OpMax) {
-            return rhs.get_max().post_link_col_type;
-        } else if (rhs.type == ExpressionInternal::exp_OpSum) {
-            return rhs.get_sum().post_link_col_type;
-        } else if (rhs.type == ExpressionInternal::exp_OpAvg) {
-            return rhs.get_avg().post_link_col_type;
-        } else if (rhs.type == ExpressionInternal::exp_OpCount) {
-            return type_Int;
-        } else if (rhs.type == ExpressionInternal::exp_OpSizeString) {
-            return type_Int;
-        } else if (rhs.type == ExpressionInternal::exp_OpSizeBinary) {
-            return type_Int;
-        }
-
-        throw std::runtime_error("Unsupported query (type undeductable). A comparison must include at lease one keypath");
-    }
-
-    template <typename LHS_T>
-    void internal_add_comparison_to_query(Query& query, LHS_T& lhs, Predicate::Comparison cmp, ExpressionContainer& rhs, DataType comparison_type)
-    {
-        switch (rhs.type) {
-            case ExpressionInternal::exp_Value:
-                do_add_comparison_to_query(query, cmp, lhs, rhs.get_value(), comparison_type);
-                return;
-            case ExpressionInternal::exp_Property:
-                do_add_comparison_to_query(query, cmp, lhs, rhs.get_property(), comparison_type);
-                return;
-            case ExpressionInternal::exp_OpMin:
-                do_add_comparison_to_query(query, cmp, lhs, rhs.get_min(), comparison_type);
-                return;
-            case ExpressionInternal::exp_OpMax:
-                do_add_comparison_to_query(query, cmp, lhs, rhs.get_max(), comparison_type);
-                return;
-            case ExpressionInternal::exp_OpSum:
-                do_add_comparison_to_query(query, cmp, lhs, rhs.get_sum(), comparison_type);
-                return;
-            case ExpressionInternal::exp_OpAvg:
-                do_add_comparison_to_query(query, cmp, lhs, rhs.get_avg(), comparison_type);
-                return;
-            case ExpressionInternal::exp_OpCount:
-                do_add_comparison_to_query(query, cmp, lhs, rhs.get_count(), comparison_type);
-                return;
-            case ExpressionInternal::exp_OpSizeString:
-                do_add_comparison_to_query(query, cmp, lhs, rhs.get_size_string(), comparison_type);
-                return;
-            case ExpressionInternal::exp_OpSizeBinary:
-                do_add_comparison_to_query(query, cmp, lhs, rhs.get_size_binary(), comparison_type);
-                return;
-        }
-    }
-
-    void add_comparison_to_query(Query &query, Predicate::Comparison cmp, ExpressionContainer& rhs)
-    {
-        DataType comparison_type = get_comparison_type(rhs);
-        switch (type) {
-            case ExpressionInternal::exp_Value:
-                internal_add_comparison_to_query(query, get_value(), cmp, rhs, comparison_type);
-                return;
-            case ExpressionInternal::exp_Property:
-                internal_add_comparison_to_query(query, get_property(), cmp, rhs, comparison_type);
-                return;
-            case ExpressionInternal::exp_OpMin:
-                internal_add_comparison_to_query(query, get_min(), cmp, rhs, comparison_type);
-                return;
-            case ExpressionInternal::exp_OpMax:
-                internal_add_comparison_to_query(query, get_max(), cmp, rhs, comparison_type);
-                return;
-            case ExpressionInternal::exp_OpSum:
-                internal_add_comparison_to_query(query, get_sum(), cmp, rhs, comparison_type);
-                return;
-            case ExpressionInternal::exp_OpAvg:
-                internal_add_comparison_to_query(query, get_avg(), cmp, rhs, comparison_type);
-                return;
-            case ExpressionInternal::exp_OpCount:
-                internal_add_comparison_to_query(query, get_count(), cmp, rhs, comparison_type);
-                return;
-            case ExpressionInternal::exp_OpSizeString:
-                internal_add_comparison_to_query(query, get_size_string(), cmp, rhs, comparison_type);
-                return;
-            case ExpressionInternal::exp_OpSizeBinary:
-                internal_add_comparison_to_query(query, get_size_binary(), cmp, rhs, comparison_type);
-                return;
-        }
-    }
-
-    bool is_null() {
-        if (type == ExpressionInternal::exp_Value) {
-            return get_value().is_null();
-        }
-        return false;
-    }
-};
+}
 
 void add_comparison_to_query(Query &query, const Predicate &pred, Arguments &args)
 {
@@ -1320,13 +511,13 @@ void add_comparison_to_query(Query &query, const Predicate &pred, Arguments &arg
     ExpressionContainer rhs(query, cmpr.expr[1], args);
 
     if (lhs.is_null()) {
-        rhs.add_null_comparison_to_query(query, cmpr);
+        add_null_comparison_to_query(query, cmpr, rhs);
     }
     else if (rhs.is_null()) {
-        lhs.add_null_comparison_to_query(query, cmpr);
+        add_null_comparison_to_query(query, cmpr, lhs);
     }
     else {
-        lhs.add_comparison_to_query(query, cmpr, rhs);
+        add_comparison_to_query(query, lhs, cmpr, rhs);
     }
 }
